+++ src/keybindings/extendAll.ts
@@ -1,4 +1,8 @@
-import { BlockUUID, ILSPluginUser } from "@logseq/libs/dist/LSPlugin";
+import {
+  BlockEntity,
+  BlockUUID,
+  ILSPluginUser,
+} from "@logseq/libs/dist/LSPlugin";
 import { debug, getCurrentBlockUUID, getSettings } from "@/common/funcs";
 
 const extend = async (blockUUID: BlockUUID | undefined) => {
@@ -12,9 +16,9 @@ const extend = async (blockUUID: BlockUUID | undefined) => {
     });
 
     if (block && block.children && block.children.length > 0) {
-      for (let item of block.children) {
-        if (Array.isArray(item) && item[0] === "uuid") {
-          await extend(item[1]);
+      for (let item of block.children as BlockEntity[]) {
+        if (item.uuid) {
+          await extend(item.uuid);
         }
       }
     }